PS-86: refactor
