+++ src/main/scala/view/util/StdButton.scala
@@ -1,4 +1,4 @@
-package view
+package view.util
 
 import scalafx.scene.control.Button
